# reservation-platform-demo-branch/backend/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Michelin Reservation Platform API
  version: 1.0.0
  description: 미슐랭 예약 플랫폼 백엔드 API 명세
servers:
  - url: http://localhost:3000/api
paths:
...
  /users/{user_id}/reservations:
    get:
      summary: 특정 유저의 예약 목록
      tags: [Reservation]
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 예약 목록 반환
    post:
      summary: 특정 유저의 예약 생성
      tags: [Reservation]
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                restaurant_id:
                  type: integer
                reservation_time:
                  type: string
                  format: date-time
                guest_count:
                  type: integer
                name:
                  type: string
      responses:
        '201':
          description: 예약 생성 성공
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 